+++ src/dagTreeItem.ts
@@ -2,16 +2,16 @@
 import * as vscode from 'vscode';
 
 export class DagTreeItem extends vscode.TreeItem {
-	public isPaused: boolean;
-	public isActive: boolean;
-	public dagId: string;
-	public owners: string[];
-	public tags: {name:string}[];
-	public apiResponse: any;
-	public fileToken: string;
-	public latestDagRunId: string;
-	public latestDagState: string;
-	public isFav: boolean = false;
+	public IsPaused: boolean;
+	public IsActive: boolean;
+	public DagId: string;
+	public Owners: string[];
+	public Tags: {name:string}[];
+	public ApiResponse: any;
+	public FileToken: string;
+	public LatestDagRunId: string;
+	public LatestDagState: string;
+	public IsFav: boolean = false;
 
 	constructor(apiResponse: any) {
 		super(apiResponse["dag_id"]);
@@ -20,42 +20,44 @@ export class DagTreeItem extends vscode.TreeItem {
 	}
 
 	public setApiResponse(apiResponse: any) {
-		this.apiResponse = apiResponse;
-		this.dagId = apiResponse["dag_id"];
-		this.isActive = apiResponse["is_active"];
-		this.isPaused = apiResponse["is_paused"];
-		this.owners = apiResponse["owners"];
-		this.tags = apiResponse["tags"];
-		this.fileToken = apiResponse["file_token"];
+		this.ApiResponse = apiResponse;
+		this.DagId = apiResponse["dag_id"];
+		this.IsActive = apiResponse["is_active"];
+		this.IsPaused = apiResponse["is_paused"];
+		this.Owners = apiResponse["owners"];
+		this.Tags = apiResponse["tags"];
+		this.FileToken = apiResponse["file_token"];
 	}
 
 	public isDagRunning(): boolean {
-		return (this.latestDagState === 'queued' || this.latestDagState === 'running');
+		return (this.LatestDagState === 'queued' || this.LatestDagState === 'running');
 	}
 
 	public refreshUI() {
-		if (this.isPaused) {
+		super.label = this.DagId;
+
+		if (this.IsPaused) {
 			this.iconPath = new vscode.ThemeIcon('circle-outline');
-			this.apiResponse["is_paused"] = true;
+			this.ApiResponse["is_paused"] = true;
 		}
 		else {
 			//"queued" "running" "success" "failed"
-			if (this.latestDagState === 'queued') {
+			if (this.LatestDagState === 'queued') {
 				this.iconPath = new vscode.ThemeIcon('loading~spin');
 			}
-			else if (this.latestDagState === 'running') {
+			else if (this.LatestDagState === 'running') {
 				this.iconPath = new vscode.ThemeIcon('loading~spin');
 			}
-			else if (this.latestDagState === 'success') {
+			else if (this.LatestDagState === 'success') {
 				this.iconPath = new vscode.ThemeIcon('check');
 			}
-			else if (this.latestDagState === 'failed') {
+			else if (this.LatestDagState === 'failed') {
 				this.iconPath = new vscode.ThemeIcon('error');
 			}
 			else {
 				this.iconPath = new vscode.ThemeIcon('circle-filled');
 			}
-			this.apiResponse["is_paused"] = false;
+			this.ApiResponse["is_paused"] = false;
 		}
 	}
 
@@ -63,13 +65,13 @@ export class DagTreeItem extends vscode.TreeItem {
 		let words: string[] = filterString.split(',');
 		let matchingWords: string[] = [];
 		for (var word of words) {
-			if (word === 'active' && !this.isPaused) { matchingWords.push(word); continue; }
-			if (word === 'paused' && this.isPaused) { matchingWords.push(word); continue; }
-			if (this.dagId.includes(word)) { matchingWords.push(word); continue; }
-			if (this.owners.includes(word)) { matchingWords.push(word); continue; }
-			if (word === 'fav' && this.isFav) { matchingWords.push(word); continue; }
+			if (word === 'active' && !this.IsPaused) { matchingWords.push(word); continue; }
+			if (word === 'paused' && this.IsPaused) { matchingWords.push(word); continue; }
+			if (this.DagId.includes(word)) { matchingWords.push(word); continue; }
+			if (this.Owners.includes(word)) { matchingWords.push(word); continue; }
+			if (word === 'fav' && this.IsFav) { matchingWords.push(word); continue; }
 
-			for(var t of this.tags)
+			for(var t of this.Tags)
 			{
 				if (t.name.includes(word)) { matchingWords.push(word); continue; }
 			}